backend/queue/requirements.txt
rq==1.15.1
redis==5.0.1
psycopg2-binary==2.9.9
requests==2.31.0
python-dotenv==1.0.0














